Refactor FormBuilderContext into 'context' folder
