+++ frontend/form-builder/src/components/ComponentSettings/RadioSettings.tsx
@@ -1,8 +1,8 @@
 import React, { useContext } from "react";
-import { FormBuilderContext } from "../../app/App";
 import { useTranslation } from "react-i18next";
 import { Textarea, Textfield } from "@digdir/design-system-react";
 import { ButtonGroup } from "./ButtonGroup";
+import { FormBuilderContext } from "../../context";
 
 export const RadioSettings = () => {
   const { currentComponent } = useContext(FormBuilderContext);
