+++ frontend/form-builder/src/components/FormBuilderSubHeader/FormBuilderSubHeader.tsx
@@ -3,8 +3,8 @@ import classes from "./FormBuilderSubHeader.module.css";
 import { Heading, Link } from "@digdir/design-system-react";
 import { FloppydiskFillIcon, TasklistSendFillIcon } from "@navikt/aksel-icons";
 import { t } from "i18next";
-import { FormBuilderContext } from "../../app/App";
 import { saveForm } from "./utils";
+import { FormBuilderContext } from "../../context";
 
 export const FormBuilderSubHeader: React.FC = () => {
   return (
